add config.h, auth.rs; reoganizacao do diretorio
