+++ auth.c
@@ -1,0 +1,32 @@
+#include <stdio.h>
+#include "config.h"
+
+int menu(void){
+    int opcao;
+
+    printf("* ##   ##   #####   ##   ##    ####   ######   ###  ### ######    # ##### *\n");
+    printf("* ###  ##  ### ###  ##   ##   ##  ##   ##  ##   ##  ##   ##  ##   ##  ##  *\n");
+    printf("* #### ##  ##   ##  ##   ##  ##        ##  ##    ####    ##  ##     ## *\n");
+    printf("* #######  ##   ##   ## ##   ##        #####      ##     #####      ## *\n");
+    printf("* ## ####  ##   ##   ## ##   ##        ## ##      ##     ##         ## *\n");
+    printf("* ##  ###  ### ###    ###     ##  ##   ## ##      ##     ##         ## *\n");
+    printf("* ##   ##   #####     ###      ####   #### ##    ####   ####       #### *\n");
+    printf("\n1. Login\n2. Cadastro\n3. Sair\n\n");
+    scanf("%d", &opcao);
+
+    switch (opcao){
+        case 1:
+            //funcao login
+            break;
+        case 2:
+            //funcao cadasstro
+            break;
+        case 3:
+            return 1;
+            break;
+        default:
+            printf("Opção invalida.\n");
+    }
+
+    return 0;
+}
