+++ novcrypt_lib/src/auth.rs
@@ -1,0 +1,9 @@
+#[no_mangle]
+pub extern "c" fn login(){
+    // Pegar user 
+    printl!("");
+    // Pegar senha usar hash
+    // Fazer consulta no banco de dados
+    // retorna 0 se estiver dados certos e existe no banco de dados
+    // senao fazer cadastro
+}
